rename cmd/args.rs to cmd/build.rs
It makes more sense to organize this module by subcommand and put the
args into the top-level mod.rs.

This file will now contain args and run functions related to twoliter
build commands.
